requirements.txt
Django==1.8.12
Pillow==3.1.1
requests
django-widget-tweaks==1.4.1
django-pagedown==0.1.1